Conflicts:
	PFA - V2/GameEngine.cpp
	PFA - V2/GameScreen.cpp
	PFA - V2/InventoryWindow.cpp
	PFA - V2/InventoryWindow.h
	PFA - V2/PFA - V2.vcxproj
	PFA - V2/PFA - V2.vcxproj.filters
	PFA - V2/Player.h

+++ PFA - V2/Camera.cpp
@@ -18,8 +18,8 @@ void Camera::update()
 
 bool Camera::moveCameraMouse(sf::Vector2i const &size)
 {
-	float dt = Singleton::getInstance()._clock->getElapsedTime().asMilliseconds();
-	float speed = 0.0001;
+	float dt = (float)Singleton::getInstance()._clock->getElapsedTime().asMilliseconds();
+	float speed = 0.0001f;
 	float calc = dt * speed;
 	bool isChangePosition = false;
 	
@@ -62,8 +62,8 @@ bool Camera::moveCameraMouse(sf::Vector2i const &size)
 
 void Camera::moveCamera(sf::Vector2i const &size)
 {
-	float dt = Singleton::getInstance()._clock->getElapsedTime().asMilliseconds();
-	float speed = 0.0001;
+	float dt = (float)Singleton::getInstance()._clock->getElapsedTime().asMilliseconds();
+	float speed = 0.0001f;
 	float calc = dt * speed;
 
 	if (this->moveCameraMouse(size) == false)
